+++ app/src/main/java/com/u1fukui/bbs/paging/thread/ThreadPagedListAdapter.kt
@@ -1,4 +1,4 @@
-package com.u1fukui.bbs.paging
+package com.u1fukui.bbs.paging.thread
 
 import android.arch.paging.PagedListAdapter
 import android.support.v7.util.DiffUtil
